Revert "Split scan_view implementation from dependencies"
This reverts commit a18042bf8aab0050b1b5d67043340336f6decfe2.
